feat: AI 챗봇/분류 백엔드 통합 및 세금 도메인 개선

See merge request s14-fintech-finance-sub1/S14P21C203!30

+++ BE/src/main/java/com/ssafy/tax7i/taxcalendar/dto/TaxDeadlineResponse.java
@@ -4,6 +4,6 @@ public record TaxDeadlineResponse(
         String taxName,
         String description,
         String deadline,
-        int dDay
+        long dDay
 ) {
 }
